Added font and additional details
Added image for Rhythm Strategy that links to the Github repo.

Added skills, roles and basic descriptions.

Created stylesheets for index, it imports the font stylesheet Roboto so the fonts can be used in setters.

Created a local folder to hold font files.

Created font faces for regular and thin to add variety.