Worked on Contact Form
Created a contactFrom.html to eventually act as iframe on main page. But want to get it to work first.

Created thankYou.html to act as a landing page after submitting form. For later.

Added style for input fields in css file to make it more presentalble. With borders that only appear on click and rounded corners.

Moved button back into the right section as it was having sizing issues.