Did more refromatting
Instead of having image elements that need to be resized to fit the div the image was made the div background so changing the div is all that is needed.

New div element was made specifically for the above.

Created a space-break element which did not work as intented.

Made the Rhythm Strategy image div a hyperlink that opens in a new page.