Fixed formatting and added mobile layout
The height of GAMES section was set to  specific pixel size to stop from filling a large portion. This does not scale well and ends up with text being cutoff if zoomed in too much.

Added a @media override if the max width is less than 420px so that objects   are shown in only 1 column.